Updating descriptions of attributes. Fixing some parsing error. Introduced unit test for building SDDCSpec and bringup E2E test.
Testing done:
make build
make lint
make test

+++ internal/resource_utils/resource_utils.go
@@ -25,7 +25,7 @@ func ToInt32Pointer(object interface{}) *int32 {
 	if object == nil {
 		return nil
 	}
-	objectAsInt32 := object.(int32)
+	objectAsInt32 := int32(object.(int))
 	return &objectAsInt32
 }
